backend/services/authorizer-service/requirements.txt
boto3
PyJWT==2.9.0
requests
pydantic==2.8.2
passlib[bcrypt]==1.7.4